+++ HISTORY.md
@@ -1,3 +1,9 @@
+v0.0.2
+==================
+* Adds Threshold Scale under the bar
+* Adds ability to hide/show the current threshold value
+* Adds ability to hide/show the threshold scale 
+
 v0.0.1
 ==================
 * Initial release